feat: render ext and builtin
